починил тесты + добавил independent provider в App.config

+++ HWT_11/DAL/Product.cs
@@ -1,17 +1,19 @@
 ﻿namespace ClassLibrary
 {
+    using System;
+
     public class Product
     {
         public string ProductName { get; set; }
 
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
 
         public int Discount { get; set; }
 
-        public double ExtendedPrice { get; set; }
+        public decimal ExtendedPrice { get; set; }
 
         public int Total { get; set; }
 
-        public int Quantity { get; set; }
+        public Int16 Quantity { get; set; }
     }
 }
